+++ packages/typescript-checker/src/plugin-tokens.ts
@@ -1,1 +1,2 @@
 export const fs = 'fs';
+export const tsCompiler = 'tsCompiler';