+++ packages/typescript-checker/testResources/single-project/src/errorInFileAbove2Mutants/todo.ts
@@ -5,7 +5,7 @@ export interface ITodo {
 }
 
 class Todo implements ITodo {
-  constructor(public name: string, public description: string, public completed: boolean) { }
+  constructor(public name: string, public description: string, public completed: boolean) {}
 }
 
 export class TodoList {
@@ -20,4 +20,3 @@ export class TodoList {
     return TodoList.allTodos;
   }
 }
-